Refactored distribution test logic
- Extracted clean up tasks for root project into separate plugin
- Migrated and refactored distributionTesting.gradle.kts into buildSrc
- Changed order for applying the plugins
