+++ buildSrc/subprojects/cleanup/cleanup.gradle.kts
@@ -1,6 +1,20 @@
+plugins {
+    `java-gradle-plugin`
+}
+
 apply { plugin("org.gradle.kotlin.kotlin-dsl")}
 
 dependencies {
-    "compile"(project(":testing"))
-    "compile"(project(":kotlinDsl"))
+    compile(project(":testing"))
+    compile(project(":kotlinDsl"))
 }
+
+gradlePlugin {
+    (plugins) {
+        "cleanup" {
+            id = "cleanup"
+            implementationClass = "org.gradle.cleanup.CleanupPlugin"
+        }
+    }
+}
+
