+++ buildSrc/subprojects/kotlin-dsl/src/main/kotlin/build-env.kt
@@ -1,0 +1,3 @@
+object BuildHost {
+    val isCiServer = System.getenv().containsKey("CI")
+}
